+++ src/test/testTransforms.ts
@@ -1,106 +1,17 @@
 /* tslint:disable:no-unused-expression */
 
 import { expect } from 'chai';
-import {
-  visit,
-  GraphQLSchema,
-  NamedTypeNode,
-  Kind,
-  GraphQLNamedType,
-  graphql,
-} from 'graphql';
-import { Request, Result } from '../Interfaces';
-import { Transform } from '../transforms/transforms';
-import { visitSchema, VisitSchemaKind } from '../transforms/visitSchema';
-import visitObject from '../transforms/visitObject';
+import { GraphQLSchema, GraphQLNamedType, graphql } from 'graphql';
 import { propertySchema, bookingSchema } from './testingSchemas';
-import makeSimpleTransformSchema from '../transforms/makeSimpleTransformSchema';
-
-function RenameTypes(renamer: (name: string) => string | undefined): Transform {
-  const reverseMap = {};
-  return {
-    transformSchema(originalSchema: GraphQLSchema): GraphQLSchema {
-      return visitSchema(originalSchema, {
-        [VisitSchemaKind.TYPE](
-          type: GraphQLNamedType,
-        ): GraphQLNamedType | undefined {
-          const newName = renamer(type.name);
-          if (newName && newName !== type.name) {
-            reverseMap[newName] = type.name;
-            const newType = Object.assign(Object.create(type), type);
-            newType.name = newName;
-            return newType;
-          }
-        },
-
-        [VisitSchemaKind.ROOT_OBJECT](type: GraphQLNamedType) {
-          return undefined;
-        },
-      });
-    },
-
-    transformRequest(originalRequest: Request): Request {
-      const newDocument = visit(originalRequest.document, {
-        [Kind.NAMED_TYPE](node: NamedTypeNode): NamedTypeNode | undefined {
-          const name = node.name.value;
-          if (name in reverseMap) {
-            return {
-              ...node,
-              name: {
-                kind: Kind.NAME,
-                value: reverseMap[name],
-              },
-            };
-          }
-        },
-      });
-      return {
-        document: newDocument,
-        variables: originalRequest.variables,
-      };
-    },
-
-    transformResult(result: Result): Result {
-      if (result.data) {
-        const newData = visitObject(result.data, (key, value) => {
-          if (key === '__typename') {
-            return renamer(value);
-          }
-        });
-        const newResult = {
-          ...result,
-          data: newData,
-        };
-        return newResult;
-      }
-      return result;
-    },
-  };
-}
-
-function FilterTypes(filter: (type: GraphQLNamedType) => Boolean) {
-  return {
-    transformSchema(schema: GraphQLSchema): GraphQLSchema {
-      return visitSchema(schema, {
-        [VisitSchemaKind.TYPE](type: GraphQLNamedType): null | undefined {
-          if (filter(type)) {
-            return undefined;
-          } else {
-            return null;
-          }
-        },
-      });
-    },
-  };
-}
+import { Transforms, makeSimpleTransformSchema } from '../transforms';
 
 describe('transforms', () => {
   describe('rename type', () => {
     let schema: GraphQLSchema;
     before(() => {
       const transforms = [
-        RenameTypes(
-          name =>
+        Transforms.RenameTypes(
+          (name: string) =>
             ({
               Property: 'House',
               Location: 'Spots',
@@ -160,7 +71,9 @@ describe('transforms', () => {
   describe('namespace', () => {
     let schema: GraphQLSchema;
     before(() => {
-      const transforms = [RenameTypes(name => `Property_${name}`)];
+      const transforms = [
+        Transforms.RenameTypes((name: string) => `Property_${name}`),
+      ];
       schema = makeSimpleTransformSchema(propertySchema, transforms);
     });
     it('should work', async () => {
@@ -210,7 +123,7 @@ describe('transforms', () => {
     let schema: GraphQLSchema;
     before(() => {
       const transforms = [
-        FilterTypes(type =>
+        Transforms.FilterTypes((type: GraphQLNamedType) =>
           ['ID', 'String', 'DateTime', 'Query', 'Booking'].includes(type.name),
         ),
       ];
